moving adp logic to separate module
