+++ packages/adp-tooling/.eslintignore
@@ -1,0 +1,2 @@
+test/fixtures
+dist
