+++ packages/adp-tooling/.eslintrc
@@ -1,0 +1,4 @@
+{
+    "extends": ["../../.eslintrc"],
+    "parserOptions": { "project": "./tsconfig.eslint.json" }
+}
