+++ packages/adp-tooling/jest.config.js
@@ -1,0 +1,2 @@
+const config = require('../../jest.base');
+module.exports = config;
