+++ packages/adp-tooling/package.json
@@ -1,23 +1,22 @@
 {
-    "name": "@sap-ux/preview-middleware",
-    "description": "Preview middleware",
+    "name": "@sap-ux/adp-tooling",
+    "description": "Tooling for UI5 adaptation projects",
     "repository": {
         "type": "git",
         "url": "https://github.com/SAP/open-ux-tools.git",
-        "directory": "packages/preview-middleware"
+        "directory": "packages/adp-tooling"
     },
     "bugs": {
-        "url": "https://github.com/SAP/open-ux-tools/issues?q=is%3Aopen+is%3Aissue+label%3Abug+label%3Apreview-middleware"
+        "url": "https://github.com/SAP/open-ux-tools/issues?q=is%3Aopen+is%3Aissue+label%3Abug+label%3Aadp-tooling"
     },
     "version": "0.0.1",
     "license": "Apache-2.0",
     "author": "@SAP/ux-tools-team",
     "main": "dist/index.js",
     "scripts": {
-        "start:fixture": "ui5 serve --config test/fixtures/simple-app/ui5.yaml",
         "build": "tsc --build",
         "watch": "tsc --watch",
-        "clean": "rimraf dist test/test-output coverage *.tsbuildinfo",
+        "clean": "rimraf --glob dist test/test-output coverage *.tsbuildinfo",
         "format": "prettier --write '**/*.{js,json,ts,yaml,yml}' --ignore-path ../../.prettierignore",
         "lint": "eslint . --ext .ts",
         "lint:fix": "eslint . --ext .ts --fix",
@@ -30,27 +29,25 @@
         "LICENSE",
         "dist",
         "ui5.yaml",
+        "bin",
         "!dist/*.map",
         "!dist/**/*.map"
     ],
     "dependencies": {
+        "@sap-ux/axios-extension": "workspace:*",
+        "@sap-ux/btp-utils": "workspace:*",
         "@sap-ux/logger": "workspace:*",
-        "@sap-ux/adp-tooling": "workspace:*",
-        "dotenv": "10.0.0",
-        "ejs": "3.1.7"
+        "@sap-ux/store": "workspace:*",
+        "prompts": "2.4.2",
+        "yazl": "2.5.1"
     },
     "devDependencies": {
         "@sap-ux/project-access": "workspace:*",
-        "@types/ejs": "3.1.0",
         "@types/express": "4.17.13",
-        "@types/prompts": "2.0.14",
-        "@types/supertest": "2.0.12",
+        "@types/yazl": "2.4.2",
         "express": "4.18.2",
         "nock": "13.2.1",
         "supertest": "6.3.2"
-    },
-    "peerDependencies": {
-        "express": "4"
     },
     "engines": {
         "pnpm": ">=6.26.1 < 7.0.0 || >=7.1.0",
