+++ packages/adp-tooling/test/fixtures/adaptation-project/package.json
