+++ packages/adp-tooling/test/fixtures/adaptation-project/ui5.yaml
