+++ packages/adp-tooling/test/fixtures/adaptation-project/webapp/i18n/i18n.properties
