+++ packages/adp-tooling/test/fixtures/adaptation-project/webapp/manifest.appdescr_variant
