--- packages/preview-middleware/src/adp/types.ts
@@ -1,9 +1,0 @@
-import type { UI5FlexLayer } from '@sap-ux/project-access';
-
-export interface DescriptorVariant {
-    layer: UI5FlexLayer;
-    reference: string;
-    id: string;
-    namespace: string;
-    content: object[];
-}
