+++ sonar-project.properties
@@ -6,7 +6,8 @@ sonar.exclusions=**/*.test.ts, **/*.test.tsx, **/*.story.tsx, **/examples/**/*, 
 sonar.cpd.exclusions=**/i18n.ts, **/fiori-elements-writer/src/packageConfig.ts, **/fiori-freestyle-writer/src/packageConfig.ts
 sonar.tests=.
 sonar.test.inclusions=**/*.test.ts, **/*.test.tsx
-sonar.javascript.lcov.reportPaths=packages/app-config-writer/coverage/lcov.info, \
+sonar.javascript.lcov.reportPaths=packages/adp-tooling/coverage/lcov.info, \
+                                packages/app-config-writer/coverage/lcov.info, \
                                 packages/axios-extension/coverage/lcov.info, \
                                 packages/backend-proxy-middleware/coverage/lcov.info, \
                                 packages/btp-utils/coverage/lcov.info, \
@@ -31,7 +32,8 @@ sonar.javascript.lcov.reportPaths=packages/app-config-writer/coverage/lcov.info,
                                 packages/ui5-proxy-middleware/coverage/lcov.info, \
                                 packages/ui5-test-writer/coverage/lcov.info, \
                                 packages/yaml/coverage/lcov.info
-sonar.testExecutionReportPaths=packages/app-config-writer/coverage/sonar-report.xml, \
+sonar.testExecutionReportPaths=packages/adp-tooling/coverage/sonar-report.xml, \
+                                packages/app-config-writer/coverage/sonar-report.xml, \
                                 packages/axios-extension/coverage/sonar-report.xml, \
                                 packages/backend-proxy-middleware/coverage/sonar-report.xml, \
                                 packages/btp-utils/coverage/sonar-report.xml, \
